refactor: share keymap conversion implementation
- this moves the duplicated keyboard conversion tables into iced_exdevtools::keymap.
- the idea is that iced_exwlshell, iced_layershell, iced_sessionlock, should share a single implementation.
- reduces duplication, also removes unused smol_str dependencies from those crates and adds winit-core where the shared helper now exists.
