+++ iced_exwlshell/Cargo.toml
@@ -36,7 +36,6 @@ window_clipboard.workspace = true
 log.workspace = true
 futures.workspace = true
 enumflags2.workspace = true
-smol_str.workspace = true
 
 # workspace features cannot be optional so we need to define the dependency here
 mundy = { version = "0.2.0", optional = true }
