+++ iced_sessionlock/Cargo.toml
@@ -34,7 +34,6 @@ sessionlockev.workspace = true
 futures.workspace = true
 window_clipboard.workspace = true
 log.workspace = true
-smol_str.workspace = true
 
 # workspace features cannot be optional so we need to define the dependency here
 mundy = { version = "0.2.0", optional = true }
